requirements.txt
numpy==1.23.5
pandas==1.5.3
python_dateutil==2.8.2
rpy2==3.5.17
streamlit==1.20.0
